requirements.txt
Flask==2.3.2
flask_sqlalchemy==3.0.3
SQLAlchemy==1.4.25
gunicorn==20.1.0
Flask_RESTful==0.3.10
python-dotenv==1.0.0
psycopg2==2.9.7
flask_apispec==0.11.4